requirements.txt
yfinance
pandas
openpyxl
anthropic
python-dotenv
streamlit
streamlit-authenticator
sqlalchemy
requests
pytest
schedule
plotly
